requirements.txt
py-cpuinfo>=9.0.0
psutil>=5.9.4
